Create slack_integration.gs
This is the simple way to integrate Google Scripts (Sheets, Docs, etc) with Slack. I've used it to push updates of Drive files to Slack teams with a description of the changes made and a link to the file. I've also used it with time-based triggers to send myself alerts and reminders.