Create screenwritingFormat.gs
This is a script that formats a Google Doc to comply with screenwriting standards for script writing. At this point, the writer has to use headings/labels to signify which element is which (Speaker, Character Intro, Action, etc.), and can then invoke the formatting function to format whatever has been written. In future versions, I might try to make it responsive in real-time like Celtx (https://www.celtx.com/index.html), but this version at least enables real-time collaborative script writing for free.