Create webappDatabase.gs
Deploying this script on a Spreadsheet as a web app turns Sheets into a database for a website. You can get the Script macro url and post form data to it through the website's HTML/JavaScript. It can log data in a table and return information to the website just like any other server.